+++ R/plot_density.r
@@ -2,6 +2,7 @@
 #'
 #' Plot density estimates for each continuous feature
 #' @param data input data
+#' @param binary_as_factor treat binary as categorical? Default is \code{TRUE}.
 #' @param geom_density_args a list of other arguments to \link{geom_density}
 #' @param title plot title
 #' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
@@ -27,15 +28,16 @@
 #' # Add color to density area
 #' plot_density(data, geom_density_args = list("fill" = "black", "alpha" = 0.6))
 
-plot_density <- function(data, geom_density_args = list(), title = NULL, ggtheme = theme_gray(), theme_config = list(), nrow = 4L, ncol = 4L, parallel = FALSE) {
+plot_density <- function(data, binary_as_factor = TRUE, geom_density_args = list(), title = NULL, ggtheme = theme_gray(), theme_config = list(), nrow = 4L, ncol = 4L, parallel = FALSE) {
 	## Declare variable first to pass R CMD check
 	variable <- value <- NULL
 	## Check if input is data.table
 	if (!is.data.table(data)) data <- data.table(data)
 	## Stop if no continuous features
-	if (split_columns(data)$num_continuous == 0) stop("No Continuous Features")
-	## Get continuous features
-	continuous <- split_columns(data)$continuous
+	split_data <- split_columns(data, binary_as_factor = binary_as_factor)
+	if (split_data$num_continuous == 0) stop("No continuous features found!")
+	## Get and reshape continuous features
+	continuous <- split_data$continuous
 	feature_names <- names(continuous)
 	dt <- suppressWarnings(melt.data.table(continuous, measure.vars = feature_names, variable.factor = FALSE))
 	## Calculate number of pages