+++ R/plot_missing.r
@@ -2,6 +2,8 @@
 #'
 #' This function returns and plots frequency of missing values for each feature.
 #' @param data input data
+#' @param group missing profile band taking a list of group name and group upper bounds. Default is \code{list("Good" = 0.05, "OK" = 0.4, "Bad" = 0.8, "Remove" = 1)}.
+#' @param geom_label_args a list of other arguments to \link{geom_label}
 #' @param title plot title
 #' @param ggtheme complete ggplot2 themes. The default is \link{theme_gray}.
 #' @param theme_config a list of configurations to be passed to \link{theme}.
@@ -12,25 +14,43 @@
 #' @seealso \link{profile_missing}
 #' @examples
 #' plot_missing(airquality)
+#' 
+#' ## Customize band
+#' plot_missing(airquality, group = list("B1" = 0, "B2" = 0.06, "B3" = 1))
+#' 
+#' ## Shrink geom_label size
+#' library(ggplot2)
+#' plot_missing(airquality, geom_label_args = list("size" = 2, "label.padding" = unit(0.1, "lines")))
 
-plot_missing <- function(data, title = NULL, ggtheme = theme_gray(), theme_config = list("legend.position" = c("bottom"))) {
-	## Declare variable first to pass R CMD check
-	pct_missing <- NULL
-	## Profile missing values
-	missing_value <- profile_missing(data)
-	## Create ggplot object
-	output <- ggplot(missing_value, aes_string(x = "feature", y = "num_missing", fill = "group")) +
-		geom_bar(stat = "identity") +
-		geom_text(aes(label = paste0(round(100 * pct_missing, 2), "%"))) +
-		scale_fill_manual("Group", values = c("Good" = "#1a9641", "OK" = "#a6d96a", "Bad" = "#fdae61", "Remove" = "#d7191c"), breaks = c("Good", "OK", "Bad", "Remove")) +
-		coord_flip() +
-		xlab("Features") + ylab("Missing Rows")
-	## Plot object
-	class(output) <- c("single", class(output))
-	plotDataExplorer(
-		plot_obj = output,
-		title = title,
-		ggtheme = ggtheme,
-		theme_config = theme_config
-	)
+plot_missing <- function(data, group = list("Good" = 0.05, "OK" = 0.4, "Bad" = 0.8, "Remove" = 1), geom_label_args = list(), title = NULL, ggtheme = theme_gray(), theme_config = list("legend.position" = c("bottom"))) {
+  ## Declare variable first to pass R CMD check
+  pct_missing <- Band <- NULL
+  ## Profile missing values
+  missing_value <- data.table(profile_missing(data))
+  ## Sort group based on value
+  group <- group[sort.list(unlist(group))]
+  invisible(lapply(seq_along(group), function(i) {
+    if (i == 1) {
+      missing_value[pct_missing <= group[[i]], Band := names(group)[i]]
+    } else {
+      missing_value[pct_missing > group[[i-1]] & pct_missing <= group[[i]], Band := names(group)[i]]
+    }
+  }))
+  ## Create ggplot object
+  output <- ggplot(missing_value, aes_string(x = "feature", y = "num_missing", fill = "Band")) +
+    geom_bar(stat = "identity") +
+    scale_fill_discrete("Band") +
+    coord_flip() +
+    xlab("Features") + ylab("Missing Rows")
+  geom_label_args_list <- list("mapping" = aes(label = paste0(round(100 * pct_missing, 2), "%")))
+  output <- output +
+    do.call("geom_label", c(geom_label_args_list, geom_label_args))
+  ## Plot object
+  class(output) <- c("single", class(output))
+  plotDataExplorer(
+    plot_obj = output,
+    title = title,
+    ggtheme = ggtheme,
+    theme_config = theme_config
+  )
 }